Wrap emitters
